requirements.txt
langchain>=0.1.0
langchain-openai>=0.0.2
langchain-ollama>=0.1.3
langchain-community>=0.0.10
langgraph>=0.0.15
faiss-cpu>=1.7.4
gradio>=4.0.0
pyyaml>=6.0
aiofiles>=23.2.1
openai>=1.3.0